Simplify CustomPathResolver for Market (#59)
Model of market is modified: Entity ProductDTORes is added as response
of the endpoint /products/productdto. Entities *Req are changed to
entities *Res in CartDTO and OrderDTO, in which they are referenced,
because they are only in the responses of the endpoints.
Updating of the output schemas of tests (json, md and xml).

In class CustomPathResolver, method getEndpointPath is needed: CartDTO
and CreditCardDTO have no endpoints and the endpoint for CartItemDTORes
must be customer/cart (not present in the model). For the rest of
entities, the endpoint is obtained from the model.