Update Market services related to distilleries, orders and ordered products (#84)
* Update market

Including OrderedProductDTO into OrderDTO

* Update tests for Carts and Orders

* Update class DistilleryRestController

find distillery by id instead of title

* Changes in PKs of Distillery and Region in SUT Market

In the schema: include the tag "x-pk" in the attributes title and name
of the entities distillery and region, respectively.

In BaseMarket: exclude entities distillery and region from using the
attribute id as uid.

* Include tests for data generation for products, distilleries and regions

* Updating test and output file names in TestMarketFuncProducts

* Updating tests and outputs for data generation for users and carts

* Update outputs of market tests due to schema changes

* Include tests grouping products in carts by distilleries and regions

* Update Order tests

* Actualizacion Market para insertar orders y ordered products

* Inserting orderedProducts in an order

* Update services to insert ordered products in an order

First, insert an empty order and its bill with cost 0.
After, one by one, insert the ordered products updating costs in the
order and the bill.

Update testMarket.json, tests and outputs related to the services.

* Update schema-testMarket.md in README.md

* Update README.md

* Changes in outputs after updating to QAGrow version 1.4.287

---------

Co-authored-by: Javier <10879637+javiertuya@users.noreply.github.com>